"efficent"
There's not that many of the 1m pixels used, so use an ordered map for
them. Now we don't need the full vector, we don't need to precalc its
size, so we can just iterate over the lines as we parse them from the
file instead of storing in memory, and do part1 and part2 in the same
pass. (To be honest, there are enough pixels used that this is probably
slower and less memory efficient.)